# original-bots/combined-command-center-archive/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DreamCo Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  active: "active",
  idle: "idle",
  error: "error",
  unknown: "unknown",
} as const;

export type BotTier = (typeof BotTier)[keyof typeof BotTier];

export const BotTier = {
  FREE: "FREE",
  PRO: "PRO",
  ENTERPRISE: "ENTERPRISE",
} as const;

/**
 * Where this bot's metadata came from
 */
export type BotSource = (typeof BotSource)[keyof typeof BotSource];

export const BotSource = {
  manifest: "manifest",
  heuristic: "heuristic",
  fallback: "fallback",
} as const;

export interface Bot {
  name: string;
  /** @nullable */
  displayName?: string | null;
  repoPath?: string;
  status: BotStatus;
  tier: BotTier;
  category: string;
  /**
   * Division slug from bot.manifest.json (e.g. DreamFinance)
   * @nullable
   */
  division?: string | null;
  capabilities?: string[];
  /** @nullable */
  entrypoint?: string | null;
  /** @nullable */
  revenueModel?: string | null;
  /** @nullable */
  owner?: string | null;
  description?: string;
  /** @nullable */
  lastHeartbeat?: string | null;
  pendingPRs?: number;
  revenue?: number;
  /** @nullable */
  lastUpdate?: string | null;
  /** Where this bot's metadata came from */
  source: BotSource;
}

export interface Repo {
  name: string;
  fullName: string;
  /** @nullable */
  description: string | null;
  isPrivate: boolean;
  url: string;
  stars: number;
  forks: number;
  updatedAt: string;
  /** @nullable */
  language?: string | null;
  openIssues?: number;
}

export interface Commit {
  sha: string;
  message: string;
  author: string;
  date: string;
  repoName: string;
  url: string;
}

export type RepoActivityRecentEventsItem = {
  type: string;
  repo: string;
  description: string;
  date: string;
};

export interface RepoActivity {
  totalCommits: number;
  totalRepos: number;
  totalBots: number;
  activeRepos: number;
  recentEvents: RepoActivityRecentEventsItem[];
}

export interface RevenueMetrics {
  totalRevenue: number;
  dailyRevenue: number;
  weeklyRevenue: number;
  monthlyRevenue: number;
  dailyTarget: number;
  weeklyTarget: number;
  monthlyTarget: number;
  currency: string;
  connected?: boolean;
}

export interface Subscription {
  id: string;
  customer: string;
  plan: string;
  status: string;
  amount: number;
  currency: string;
  currentPeriodEnd: string;
}

export interface Transaction {
  id: string;
  amount: number;
  currency: string;
  status: string;
  description: string;
  date: string;
}

export interface BuddyMessageInput {
  message: string;
  sessionId?: string;
}

export interface BuddyResponse {
  message: string;
  sessionId: string;
  timestamp: string;
  /** @nullable */
  emotion?: string | null;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  buddy: "buddy",
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  /** @nullable */
  emotion?: string | null;
}

export interface DashboardSummary {
  totalBots: number;
  activeBots: number;
  idleBots: number;
  totalRepos: number;
  dailyRevenue: number;
  monthlyRevenue: number;
  dailyTarget: number;
  monthlyTarget: number;
  totalSubscriptions: number;
  recentCommits: number;
  profitTargetDaily?: number;
  profitTargetMonthly?: number;
}

export interface BuildProgressSection {
  name: string;
  done: number;
  total: number;
}

export type BuildProgressDetail = { [key: string]: unknown };

export interface BuildProgress {
  overallPercent: number;
  sections: BuildProgressSection[];
  detail?: BuildProgressDetail;
}

export interface BotRunResult {
  ok: boolean;
  name: string;
  status: string;
  invocations: number;
  lastHeartbeat: string;
  message?: string;
}

export interface CopilotPr {
  number: number;
  title: string;
  url: string;
  author: string;
  createdAt: string;
  /** @nullable */
  mergeable: boolean | null;
  /** @nullable */
  mergeableState?: string | null;
  /** True when PR touches exactly one bots/* dir with a valid manifest */
  contractCompliant: boolean;
  complianceReason?: string;
  touchedDirs: string[];
  manifestPresent?: boolean;
}

export interface CopilotRewriteResponse {
  issueNumber: number;
  issueUrl: string;
}

export interface Tier {
  name: string;
  priceMonthly: number;
  requestsPerMonth: number;
  concurrentBots: number;
  models: string[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};