# original-bots/combined-command-center-archive/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DreamCo Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots/{name}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bots/{name}/run:
    post:
      operationId: runBot
      tags: [bots]
      summary: Trigger a manual run of a bot (records heartbeat + invocation)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot run recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotRunResult"
        "404":
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Bot:
      type: object
      required: [name, status, tier, category, source]
      properties:
        name:
          type: string
        displayName:
          type: ["string", "null"]
        repoPath:
          type: string
        status:
          type: string
          enum: [active, idle, error, unknown]
        tier:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
        category:
          type: string
        division:
          type: ["string", "null"]
          description: Division slug from bot.manifest.json (e.g. DreamFinance)
        capabilities:
          type: array
          items:
            type: string
        entrypoint:
          type: ["string", "null"]
        revenueModel:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        description:
          type: string
        lastHeartbeat:
          type: ["string", "null"]
        pendingPRs:
          type: integer
        revenue:
          type: number
        lastUpdate:
          type: ["string", "null"]
        source:
          type: string
          enum: [manifest, heuristic, fallback]
          description: Where this bot's metadata came from
...
    BotRunResult:
      type: object
      required: [ok, name, status, invocations, lastHeartbeat]
      properties:
        ok:
          type: boolean
        name:
          type: string
        status:
          type: string
        invocations:
          type: integer
        lastHeartbeat:
          type: string
        message:
          type: string
...